Add RemoteOutputServiceDirectory
This commit adds a new type that implements both a gRPC server for the
RemoteOutputService protocol, and a FUSE directory. It acts as the
top-level directory for all builds managed by bb_clientd.

For each Bazel workspace, it creates a PrepopulatedDirectory through
StartBuild(). The BatchCreate() and BatchStat() RPCs may be used to make
bulk changes to the contents of the PrepopulatedDirectory. Because these
RPCs use pathname strings, it uses the newly created path resolver
package to convert these request into traversals against the directory
objects.

The FinalizeBuild() call can be used to signal completion of the current
build. This currently doesn't do anything, besides removing the build ID
from the bookkeeping, so that future RPCs can no longer be received. In
the future we could add some logic here to snapshot builds.

The Clean() call can be used to remove all data associated with a single
build, similar to a 'rm -rf', but done through a single RPC. This call
significantly speeds up invocations of 'bazel clean'.