Add retrying of failed FUSE file system operations
The Bazel OutputService interface doesn't allow us to propagate I/O
errors triggered through the FUSE file system. This is annoying, because
it means that flakiness leads to build failures that could easily be
prevented by doing retries.

This change adds an ErrorRetryingBlobAccess that acts as a decorator for
BlobAccess, retrying failed reads against the CAS. We make sure that all
CAS-backed files are created using this BlobAccess. This should make
bb_clientd more robust, especially for CI workloads.