Add the ability to upload all output path contents into the CAS
Right now bb_clientd is only capable of restoring files in the output
path across restarts if they are also present remotely. No local copies
are preserved. This change adds the ability to restore the entire output
path by making sure that at the end of a build, we upload everything
that is not present remotely.

To achieve this we use the already existing BatchedStoreBlobAccess
that's part of bb-remote-execution. It is capable of uploading blobs if
and only if they are not present remotely.