Let the Remote Output Service expose last modified times
It turns out that Bazel depends on having this information present for
bb_clientd. For regular files, Bazel doesn't really care about having
file timestamps, as it internally just uses the hash and size. For
directories it does depend on knowing timestamps. This can be seen by
looking at ActionMetadataHandler.fileArtifactValueFromStat(). For
directories, it calls FileArtifactValue.createForDirectoryWithMtime().

This change upgrades to a newer version of bb-remote-execution, which
includes an implementation of InMemoryPrepopulatedDirectory that tracks
this information as an attribute. It then alters the Remote Output
Service implementation to return this attribute as part of BatchStat.