Reconfigure bb_clientd to also act as a system local cache
When building projects without remote execution, it may sometimes be
desirable to have a system local cache. Bazel already offers a feature
like this through its --disk_cache flag, but some limitations it has
continue to remain unaddressed.

The advantage of offering this as part of bb_clientd is that storage
space is shared with the cache that's used for proxying. The use of
LocalBlobAccess also means that data is stored more densely, and that
the space that is used is bounded. There is no need to do manual
cleanups.