Upgrade all Go dependencies
This requires us to catch up with changes to bb-storage and
bb-remote-execution. There we've added BlobAccess.GetFromComposite(),
and have overhauled DirectoryFetcher as a result of that.

GlobalTreeContext (the code behind ~/bb_clientd/cas/*/tree/*/) relied on
being able to address child directories in a Tree in an iterative
manner using IndexedTree, but this is no longer possible. Child
directories now need to be addressed by parent and child digest, as
that's how GetFromComposite() is invoked. This means that NFSv4 file
handles for those directories now need to contain SHA sums for both.
Fortunately, 128 bytes is still large enough to accommodate that.

One advantage of letting file handles contain both SHA sums is that
resolution of children is stateless, allowing us to eliminate
DirectoryContentsContext. LookupDirectory() can also no longer fail,
which is nice. This is also how things looked at some point in the past
when we only supported FUSE.