Add the digest function to the pathname scheme
As SHA256TREE hashes are the same length as SHA-256, we need to add
something to pathnames to be able to distinguish between the two. Let's
solve that by adding one directory per hash function. This means that
any path in bb_clientd will now have "sha256" in its name.