Let maximum_state_file_age work in epochs
bb_clientd is capable of periodically ignoring state files, meaning that
it periodically does an automatic 'bazel clean'. This is great, as it
prevents unused files from accumulating in the central CAS indefinitely.

What is slightly annoying is that this logic is exercised by each
bb_clientd instance independently. This means that if you run a build
with a larger number of stages, there is a higher probability that the
build is slowed down by a single worker discarding its state.

Let's reduce the likelihood of this happening by letting bb_clientd
ignore the state files through a given cadence. This means that if one
of the legs is the build of slow, all of them well. Or the inverse: all
of them will be fast.